+++ Sources/GitHub/Issue.swift
@@ -1,9 +1,11 @@
 import Foundation
 
-public class Issue: Decodable {
+public struct Issue {
 	public var url: URL
 	public var number: Int
+}
 
+extension Issue: Decodable {
 	enum CodingKeys: String, CodingKey {
 		case url = "html_url"
 		case number
