+++ Sources/GitHub/PullRequest.swift
@@ -1,9 +1,11 @@
 import Foundation
 
-public class PullRequest: Decodable {
+public struct PullRequest {
 	public var url: URL
 	public var number: Int
+}
 
+extension PullRequest: Decodable {
 	enum CodingKeys: String, CodingKey {
 		case url = "html_url"
 		case number
